long > integer

+++ src/main/java/cz/jpalcut/pia/dao/TemplateDAO.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TemplateDAO extends JpaRepository<Template, Long> {
+public interface TemplateDAO extends JpaRepository<Template, Integer> {
 
 }
